refactor: SecurityContextHolderFilter 패키지를 context로 이동
