+++ src/main/java/nextstep/authentication/context/SecurityContextHolderFilter.java
@@ -1,4 +1,4 @@
-package nextstep.authentication.filter;
+package nextstep.authentication.context;
 
 import jakarta.servlet.Filter;
 import jakarta.servlet.FilterChain;
@@ -7,8 +7,6 @@ import jakarta.servlet.ServletRequest;
 import jakarta.servlet.ServletResponse;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import nextstep.authentication.context.SecurityContextHolder;
-import nextstep.authentication.context.SecurityContextRepository;
 
 import java.io.IOException;
